jf: two-way blobs (2/2)
TARGET_USES_GPE_RIL will handle everything. Move system properties in a makefile to handle rild library path which is different for both methods. Loki is common for both, so move it in a dedicated folder. Releasetools is method-based so we'll handle in two ways as well (normal and alternate, this last for GPE RIL).