AudioSystemLegacy.h: Support new audio devices enum values introduced in JB MR1 code drop
Today, each audio device is identified by a bit in a 32 bit field with
16 bits for output devices and 16 bits for input devices.
So the number of output devices is now exhausted i.e. we can't add new output devices.

So in order to add new devices, a new convention for audio device
identification is defined, allowing 31 different input and output devices.
Bit 31 of the bit field is reserved and when set indicates an input device.

Change-Id: Ib9b56e63ae70a02bca8a038d14a1bbd281883e6a
Signed-off-by: ndifolcx <neilx.di.folco@intel.com>
Signed-off-by: Jin Wei <wei.a.jin@intel.com>

Original commit: https://android-review.googlesource.com/#/c/92788/